+++ microauth_fastapi/__init__.py
@@ -28,7 +28,7 @@ from .exceptions import (
 from .models import Customer, PlatformMonthlyAllowance
 from .sdk import MicroAuth
 
-__version__ = "2.6.0"
+__version__ = "2.7.0"
 
 __all__ = [
     "AuthDenied",